refactor(task-status-sequencer): architectural improvements, purity & separation of concerns
