+++ src/features/task-status-sequencer/app/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * App constants: thresholds and timing.
+ */
+
+/** Default long-press threshold (ms) for cancel intent. */
+export const LONG_PRESS_CANCEL_MS = 500;
